last commit

+++ sb103/src/main/java/com/kgisl/sb103/entity/Person.java
@@ -2,5 +2,6 @@ package com.kgisl.sb103.entity;
  
 import org.springframework.data.annotation.Id;
  
-public record Person(@Id Long p_id, String p_name, String p_email) {
+public record Person(@Id int p_id, String p_name, String p_email) {
 }
+
